Add new options "Email Leads", "Email Requester", and "Password Changed"
The email options replace the Hold option and bring up a text box with a
message for the performer to edit and send. In addition the request status
is changed and the action is logged.

The password option is meant to be used when an account request is for an
existing account and is really a password reset request.  This option should
be chosed after the password is changed and results in the request status being
changed to Approved and the password change being logged.